+++ backend/tests/component/core/constraint_validators/test_attribute_numberpool_constraints.py
@@ -18,7 +18,7 @@ from infrahub.core.validators.attribute.number_pool import (
 from infrahub.core.validators.enum import ConstraintIdentifier
 from infrahub.core.validators.model import SchemaConstraintValidatorRequest
 from infrahub.database import InfrahubDatabase
-from infrahub.pools.tasks import SchemaNumberPoolValidator
+from infrahub.pools.schema_number_pool_synchronizer import SchemaNumberPoolSynchronizer
 from tests.helpers.schema.snow import SNOW_INCIDENT, SNOW_REQUEST, SNOW_TASK
 
 
@@ -28,7 +28,7 @@ async def snow_incident_01(db: InfrahubDatabase, default_branch: Branch, registe
     registry.schema.register_schema(schema=schema, branch=default_branch.name)
     registry.node[InfrahubKind.NUMBERPOOL] = CoreNumberPool
 
-    snpv = SchemaNumberPoolValidator(db=db, log=MagicMock(), schema_manager=registry.schema)
+    snpv = SchemaNumberPoolSynchronizer(db=db, log=MagicMock(), schema_manager=registry.schema)
     await snpv.run()
 
     incident_1 = await Node.init(db=db, schema="SnowIncident", branch=default_branch)
