+++ backend/tests/functional/convert_object_type/test_convert_object_type.py
@@ -10,7 +10,7 @@ from infrahub.core.constants.infrahubkind import NUMBERPOOL
 from infrahub.core.query.resource_manager import NumberPoolGetReserved
 from infrahub.core.registry import registry
 from infrahub.core.schema import AttributeSchema, GenericSchema, NodeSchema, SchemaRoot
-from infrahub.pools.tasks import SchemaNumberPoolValidator
+from infrahub.pools.schema_number_pool_synchronizer import SchemaNumberPoolSynchronizer
 from tests.helpers.test_app import TestInfrahubApp
 
 if TYPE_CHECKING:
@@ -158,7 +158,7 @@ class TestConvertObjectType(TestInfrahubApp):
 
 class TestConvertObjectTypeResourcePool(TestInfrahubApp):
     async def _run_number_pool_validator(self, db: InfrahubDatabase) -> None:
-        snpv = SchemaNumberPoolValidator(db=db, log=MagicMock(), schema_manager=registry.schema)
+        snpv = SchemaNumberPoolSynchronizer(db=db, log=MagicMock(), schema_manager=registry.schema)
         await snpv.run()
 
     @pytest.fixture
